Extract CactusBoxPtr and CactusBox into a ptr module
Rename to RcBoxPtr and RcBox
